Extract character_stat to separate header
